fix(recipes): preserve count expression contracts
Keep list-size declarations at their original integer type, reject multi-ID activity filters, and split the count validation checklist into independent rules.\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>

+++ code-conversion/recipes/src/test/java/io/camunda/migration/code/recipes/delegate/JavaDelegateWithClientQueryRecipesTest.java
@@ -114,7 +114,7 @@ public class JavaDelegateWithClientQueryRecipesTest implements RewriteTest {
                             .page()
                             .totalItems();
 
-                    Long pagedCount = camundaClient
+                    int pagedCount = camundaClient
                             .newProcessInstanceSearchRequest()
                             .filter(filter -> filter
                                     .processDefinitionId("example-workflow-process")
@@ -122,7 +122,7 @@ public class JavaDelegateWithClientQueryRecipesTest implements RewriteTest {
                             .send()
                             .join()
                             .page()
-                            .totalItems();
+                            .totalItems().intValue();
 
                     Long streamCount = camundaClient
                             .newProcessInstanceSearchRequest()
